backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "3.12.7"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "pydantic[email] (>=2.11.4,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "sqlalchemy[asyncio] (>=2.0.40,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "fastapi-users[sqlalchemy] (>=14.0.1,<15.0.0)",
    "orjson (>=3.10.18,<4.0.0)",
    "minio-async (>=1.0.1,<2.0.0)",
    "minio (>=7.2.15,<8.0.0)",
]


[tool.poetry.group.testing.dependencies]
httpx = "^0.28.1"
pytest-asyncio = "^0.26.0"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "session"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
